Make ball bounce in left and right direction correctly along all all top edge of paddle. Closer to the edge the ball contacts the paddle the steeper the trajectory (vector angle). Closer to the middle, straighter up the ball bounces
TODO
Make invisible boundary to enable brick-collision check
to save thousands of for loop iterations on each frame that is
not in the boundary. So that in empty space where there will
never be bricks, then there is not brick for loop check

TODO
Bonus points given for hitting balls close to the middle of bricks. Visual
cue like color change on brick or something.